# documentation/api/openapi.yaml
openapi: 3.1.1
info:
  title: Homebank API
  version: 1.0.0
  description: |
    Machine-readable API contract for Homebank backend.
    Human-readable companion docs live in documentation/api/endpoints.md.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /customers/transactionHeads/{transactionHeadId}:
    get:
      tags: [Customers]
      summary: Get all customers and one transaction head
      parameters:
        - $ref: '#/components/parameters/transactionHeadId'
      responses:
        '200':
          description: Composite returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomersAndTransactionHeadDTO'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /customers/{customerId}/transactionHeads/{transactionHeadId}:
    get:
      tags: [Customers]
      summary: Get customer and transaction head
      parameters:
        - $ref: '#/components/parameters/customerId'
        - $ref: '#/components/parameters/transactionHeadId'
      responses:
        '200':
          description: Composite returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerAndTransactionHeadDTO'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /transactionHeads/{transactionHeadId}:
    parameters:
      - $ref: '#/components/parameters/transactionHeadId'
    get:
      tags: [Transaction Heads]
      summary: Read a transaction head accessible through either customer
      responses:
        '200':
          description: Transaction head
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionHeadDTO'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags: [Transaction Heads]
      summary: Update a transaction head accessible through either stored customer
      description: The path supplies the ID. Lender and borrower cannot be changed after creation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransactionHeadDTO'
      responses:
        '200':
          description: Transaction head updated
          content:
            application/json:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'

    delete:
      tags: [Transaction Heads]
      summary: Mark transaction head as deleted
      description: Requires access through either stored lender or borrower. The path identifies the head.
      parameters:
        - name: rowVersion
          in: query
          required: true
          description: The row version returned when reading the head, including any fractional seconds.
          schema:
            type: string
            format: date-time
          example: '2026-01-01T00:00:00'
      responses:
        '204':
          description: Transaction head deleted
        '400':
          $ref: '#/components/responses/BadRequestError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    customerId:
      name: customerId
      in: path
      required: true
      schema:
        type: integer
        format: int32
    transactionHeadId:
      name: transactionHeadId
      in: path
      required: true
      schema:
        type: integer
        format: int32
...
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            validationFailed:
              summary: Validation failed
              value:
                timestamp: "2026-03-14T10:00:00Z"
                status: 400
                error: "Bad Request"
                code: "VALIDATION_FAILED"
                message: "Validation failed for one or more fields."
                path: "/auth/register"
                details:
                  - code: "NOT_BLANK"
                    field: "email"
                    message: "Email is missing"
...
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            accountNotActivated:
              summary: Pending account blocked by UserStatusFilter
              value:
                timestamp: "2026-03-14T10:00:00Z"
                status: 403
                error: "Forbidden"
                code: "ACCOUNT_NOT_ACTIVATED"
                message: "Please activate your account first."
                path: "/customers/1"
                details: null
            resourceAccessDenied:
              summary: Resource authorization failed
              value:
                timestamp: "2026-03-14T10:00:00Z"
                status: 403
                error: "Forbidden"
                code: "RESOURCE_ACCESS_DENIED"
                message: "You do not have permission to access this customer."
                path: "/customers/1"
                details:
                  resourceType: "CUSTOMER"
                  resourceId: 1
                  action: "READ"
                  metadata:
                    userId: 42
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            resourceNotFound:
              summary: Requested resource does not exist
              value:
                timestamp: "2026-03-14T10:00:00Z"
                status: 404
                error: "Not Found"
                code: "RESOURCE_NOT_FOUND"
                message: "The requested resource was not found."
                path: "/customers/999"
                details:
                  resourceType: "CUSTOMER"
                  resourceId: 999
                  metadata: null
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            rowVersionMismatch:
              summary: Optimistic lock conflict
              value:
                timestamp: "2026-03-14T10:00:00Z"
                status: 409
                error: "Conflict"
                code: "ROW_VERSION_MISMATCH"
                message: "The resource was modified by another process. Please refresh and try again."
                path: "/customers/1"
                details: null
...
  schemas:
    ApiError:
      type: object
      required: [timestamp, status, error, code, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        code:
          type: string
        message:
          type: string
        path:
          type: string
        details:
          description: Optional structured context. Shape depends on `code`.
          nullable: true
...
    CustomerDTO:
      type: object
      required: [id, name, typeOfCustomerCode, customerAmount, rowVersion]
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
          nullable: true
        typeOfCustomerCode:
          type: string
        customerAmount:
          type: integer
          format: int32
        rowVersion:
          type: string
          format: date-time
...
    UpdateTransactionHeadDTO:
      type: object
      required: [transactionName, startDate, rowVersion]
      properties:
        transactionName:
          type: string
        description:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
        prelEndDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        rowVersion:
          type: string
          format: date-time
    TransactionHeadDTO:
      type: object
      required: [id, lenderId, borrowerId, transactionName, startDate]
      properties:
        id:
          type: integer
          format: int32
        lenderId:
          type: integer
          format: int32
        borrowerId:
          type: integer
          format: int32
        transactionName:
          type: string
        description:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
        prelEndDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        amount:
          type: integer
          format: int32
          minimum: 0
          nullable: true
        borrower:
          type: string
          nullable: true
        lender:
          type: string
          nullable: true
        rowVersion:
          type: string
          format: date-time
          nullable: true
...
    CustomersAndTransactionHeadDTO:
      type: object
      required: [customers, transactionHead]
      properties:
        customers:
          type: array
          items:
            $ref: '#/components/schemas/CustomerDTO'
        transactionHead:
          $ref: '#/components/schemas/TransactionHeadDTO'
...
    CustomerAndTransactionHeadDTO:
      type: object
      required: [customer, transactionHead]
      properties:
        customer:
          $ref: '#/components/schemas/CustomerDTO'
        transactionHead:
          $ref: '#/components/schemas/TransactionHeadDTO'
...